fix(logging): Rename logging package to log_config
Fixes name collision with Python's built-in logging module that was
causing import errors in CI. The collision occurred when Python tried
to import 'logging.handlers' but found our local clab_tools/logging
directory instead of the standard library.

Changes:
- Rename clab_tools/logging/ to clab_tools/log_config/
- Update all imports throughout the codebase
- Update documentation examples
- All tests pass locally

Resolves CI error: ModuleNotFoundError: No module named 'logging.handlers'
